Added direct property to Button.js
I noticed one of the comments on your youtube video pointed out that every button made using Button.js redirects to the sign-up page. I've added a direct property, along with a checkDirect function that will update the path of the button if one is passed through, otherwise the direct property defaults to the sign-up page. Hopefully this makes Button.js more reusable.